requirements.txt
asgiref==3.9.1
Django==5.2.4
django-extensions==4.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
dotenv~=0.9.9
python-dotenv~=1.1.1
djangorestframework~=3.16.1
Faker~=37.5.3
django-filter~=25.1
django-debug-toolbar~=6.0.0
celery~=5.5.3
gunicorn
drf-spectacular~=0.27.2
djangorestframework-simplejwt~=5.3.1
redis~=6.4.0
django-redis
black~=25.1.0
ruff~=0.12.9